pgdg-autopkgtest: Stop trying to install all binaries manually
So far we tried to manually install all binaries for a source package so
we don't have to rebuild the source once a new PG version is added to
the "pgdg" default set. Unfortunately this needs a gross list of
exceptions, and utterly fails for postgresql-multicorn because its
binaries conflict with each other. So give up that plan, and let
autopkgtest handle installing binaries, except when
PG_SUPPORTED_VERSIONS is set to a specific version.